+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+# v0.0.140 - 2019-08-12
+- Update icons to match the 1.37.0 GUI (great work by [johnletey](https://github.com/johnletey) - many thanks!)
 - Add option to hide ruler markers
 - Add option to reveal tree when clicking the status bar